+++ src/mantle/testing/class-test-case.php
@@ -44,6 +44,8 @@ use function Mantle\Support\Helpers\collect;
  * Root Test Case for Mantle sites.
  *
  * Not designed for external use. Use {@see Mantle\Testkit\Test_Case} instead.
+ *
+ * @property-read Application|null $app
  */
 abstract class Test_Case extends BaseTestCase {
 	use Assertions,
@@ -97,7 +99,6 @@ abstract class Test_Case extends BaseTestCase {
 		static::register_traits();
 
 		if ( ! empty( static::$test_uses ) ) {
-
 			static::get_test_case_traits()
 				->each(
 					function( $trait ) {
@@ -266,7 +267,7 @@ abstract class Test_Case extends BaseTestCase {
 	/**
 	 * Get an array of priority traits.
 	 *
-	 * @return array
+	 * @return array<class-string>
 	 */
 	protected static function get_priority_traits(): array {
 		return [
@@ -280,14 +281,14 @@ abstract class Test_Case extends BaseTestCase {
 	/**
 	 * Register the traits that this test case uses.
 	 */
-	public static function register_traits() {
+	public static function register_traits(): void {
 		static::$test_uses = array_flip( class_uses_recursive( static::class ) );
 	}
 
 	/**
 	 * Refresh the application instance.
 	 */
-	protected function refresh_application() {
+	protected function refresh_application(): void {
 		$this->app = $this->create_application();
 
 		if ( class_exists( Facade::class ) ) {
@@ -316,24 +317,26 @@ abstract class Test_Case extends BaseTestCase {
 	}
 
 	/**
-	 * Allow the factory to be checked against.
+	 * Allow the factory/app to be checked against.
 	 *
 	 * @param string $name Property name.
 	 * @return boolean
 	 */
-	public function __isset( $name ) {
-		return 'factory' === $name;
+	public function __isset( $name ): bool {
+		return 'factory' === $name || 'app' === $name;
 	}
 
 	/**
-	 * Retrieve the factory instance non-statically.
+	 * Retrieve the factory/app instance non-statically.
 	 *
 	 * @param string $name Property name.
 	 * @return mixed
 	 */
 	public function __get( $name ) {
-		if ( 'factory' === $name ) {
-			return self::factory();
-		}
+		return match ( $name ) {
+			'factory' => self::factory(),
+			'app' => $this->app,
+			default => null,
+		};
 	}
 }